+++ packages/ui-components/helpers/AppProviders.js
@@ -8,19 +8,27 @@ import { MuiThemeProvider, StylesProvider } from '@material-ui/core/styles';
 import { ThemeProvider } from 'styled-components';
 import CssBaseline from '@material-ui/core/CssBaseline';
 import PropTypes from 'prop-types';
-import theme from '../stories/story-utils/theme';
+import { lightTheme, darkTheme } from '../stories/story-utils/theme';
 
-export const AppProviders = ({ children }) => (
-  <StylesProvider injectFirst>
-    <MuiThemeProvider theme={theme}>
-      <ThemeProvider theme={theme}>
-        <CssBaseline />
-        {children}
-      </ThemeProvider>
-    </MuiThemeProvider>
-  </StylesProvider>
-);
+export const AppProviders = ({ params, children }) => {
+  const theme = params?.theme === 'dark' ? darkTheme : lightTheme;
+  return (
+    <StylesProvider injectFirst>
+      <MuiThemeProvider theme={theme}>
+        <ThemeProvider theme={theme}>
+          <CssBaseline />
+          {children}
+        </ThemeProvider>
+      </MuiThemeProvider>
+    </StylesProvider>
+  );
+};
 
 AppProviders.propTypes = {
+  params: PropTypes.object,
   children: PropTypes.any.isRequired,
+};
+
+AppProviders.defaultProps = {
+  params: null,
 };
